% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundRobin.R
\docType{data}
\name{AFLW}
\alias{AFLW}
\title{AFLW}
\format{
A character string vector with 18 team names
}
\usage{
AFLW
}
\description{
Australian Women's Football League teams for the national competition
}
\keyword{datasets}
