% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundRobin.R
\name{randomize}
\alias{randomize}
\alias{randomise}
\alias{randomize.default}
\title{Arrange the components of an object in random order}
\usage{
randomize(object, ...)

randomise(object, ...)

\method{randomize}{default}(object, ...)
}
\arguments{
\item{object}{Any object with ordered components}

\item{...}{Additional arguments to be passed on to methods}
}
\value{
A similar object with the components arranged in random
        order, but with the names and other attributes held fixed.
}
\description{
Simple utility to permute the components of an object
}
\examples{
Season_2021 <- randomise(round_robin(AFLW))
summary(Season_2021, "travel")

}
