% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundrobin.R
\name{roundrobin}
\alias{roundrobin}
\title{Transform a data.frame or tibble into a wide form of a round-robin combination of a particular key column.}
\usage{
roundrobin(data, key)
}
\arguments{
\item{data}{target data.frame or tibble}

\item{key}{target key for combination calcuration}
}
\description{
Transform a data.frame or tibble into a wide form of a round-robin combination of a particular key column.
}
